requirements.txt
'''
requirements.txt
'''
streamlit==1.44.1
openai==1.76.0
google-generativeai>=0.8.5
pytesseract>=0.3.13
Pillow>=11.2.1
PyMuPDF==1.25.5
python-dotenv==1.1.0
